fix int types to use QT types, it seems to be important small changes to directorypicker for bug PCLOUD-397

+++ pCloud/psettings.cpp
@@ -64,10 +64,10 @@ void PSettings::seti(const QString &key, int val){
     settings->setValue(key, val);
 }
 
-uint64_t PSettings::getu64(const QString &key){
+quint64 PSettings::getu64(const QString &key){
     return settings->value(key).toULongLong();
 }
 
-void PSettings::setu64(const QString &key, uint64_t val){
-    settings->setValue(key, (unsigned long long)val);
+void PSettings::setu64(const QString &key, quint64 val){
+    settings->setValue(key, val);
 }
